getTypeByRange input Long

+++ spark-gbm/src/main/scala/org/apache/spark/ml/gbm/util/Utils.scala
@@ -49,7 +49,7 @@ private[gbm] object Utils extends Logging {
   /**
     * Choose the most compact Integer Type according to the range
     */
-  def getTypeByRange(range: Int): String = {
+  def getTypeByRange(range: Long): String = {
     require(range >= 0)
 
     if (range <= Byte.MaxValue) {
